create-genesis-and-keys: Set DPARAM to 1 at genesis.
We can't set up the stake pools until after we're in Shelley, but
decentralisationParam takes effect immediately. This results in the
initial Shelley epochs having too few blocks, breaking invariants.

New testnets should wait until stake pools are up and running and then
update the decentralisationParam with an update proposal (perhaps at
the same time as the move to Allegra)